Add ItemStackEditorResult type and update imports in related files
