+++ src/editor/editors/ItemStackEditor/index.ts
@@ -1,1 +1,2 @@
-export {default as ItemStackEditor} from './ItemStackEditor'
+export {default as ItemStackEditor} from './ItemStackEditor';
+export type { ItemStackEditorResult } from './ItemStackEditor';
